+++ src/main/java/com/cotato/networking1/controller/PropertyController.java
@@ -8,7 +8,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.HashMap;
 import java.util.List;
@@ -16,34 +15,24 @@ import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api")
+@RequestMapping("/api/properties")
 public class PropertyController {
 
     private final PropertyService propertyService;
 
-    @PostMapping("/test-data")
-    public ResponseEntity<String> uploadFile(@RequestParam("file") MultipartFile file) {
-        try {
-            propertyService.saveProperties(file);
-            return ResponseEntity.ok("File uploaded and data saved successfully");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Could not upload the file: " + e.getMessage());
-        }
-    }
-
-    @GetMapping("/properties")
+    @GetMapping()
     public ResponseEntity<PropertyListResponse> getPropertiesByZipCode(@RequestParam("zip-code") String zipCode) {
         return ResponseEntity.ok(propertyService.findPropertiesByZipCode(zipCode));
     }
 
-    @PostMapping("/properties")
+    @PostMapping()
     public ResponseEntity<?> createProperty(@RequestBody PropertyPostRequestDto propertyPostRequestDto) {
         Property property = Property.toProperty(propertyPostRequestDto);
         Property savedProperty = propertyService.saveProperty(property);
         return ResponseEntity.ok().body(Map.of("id", savedProperty.getId()));
     }
 
-    @DeleteMapping("/properties")
+    @DeleteMapping()
     public ResponseEntity<?> deletePropertyByRoadNameAddress(@RequestParam("road-name-address") String roadNameAddress) {
         try {
             List<Long> deletedIds = propertyService.deletePropertiesByRoadAddress(roadNameAddress);
